Removed _PG_INIT
Loading session variables in _PG_INIT sometimes caused access-rights
problems on the session_varible.variables table because "security
definer" cannot be specified for it. In _PG_INIT invoking
session_variable.init() via the SPI caused problems when invoked from a
_PG_INIT function in a preload library caused problems because the
catalog cache is not fully filled at that moment. So a boolean is added
to check if the variable table is already loaded in each external
function. These functions run as "security definer", so no access rights
are violated in any case.