Added functions get_stable() and get_constant()
The get() function is marked VOLATILE. And it must be so as the content
ov a variable may change during the execution of a statement, for
example in a trigger.

But in most cases the value will not change, so a function marked
"STABLE" may allow the optimizer to cache results might prove
beneficial.